move v2 schemas
